all: Move bootstrap command to flynn-host
