--- bootstrap/bootstrapper/.gitignore
@@ -1,1 +1,0 @@
-bootstrapper
