+++ script/build-deb
@@ -53,7 +53,6 @@ main() {
     --config-files /etc/flynn/version.json \
     --config-files /etc/init/flynn-host.conf \
     pinkerton/pinkerton=/usr/bin/ \
-    bootstrap/bin/flynn-bootstrap=/usr/bin/ \
     host/bin/flynn-host=/usr/bin/ \
     host/bin/flynn-init=/usr/bin/ \
     cli/flynn-cli=/usr/bin/flynn \
